pyproject.toml
[project]
name = "dynamo-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.39",
    "dotenv>=0.9.9",
    "duckdb>=1.4.0",
    "psycopg2-binary>=2.9.10",
    "typer[all]",
]
